Move FitVal to its own file and update serialization
